refactor(nexus-web): Organization settings UI module and mobile list-detail
