+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/app/organizations/[orgId]/settings/general/page.tsx
@@ -1,0 +1,1 @@
+export { default } from './general';
