+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/app/organizations/[orgId]/settings/workspaces/workspaces.css
@@ -1,0 +1,7 @@
+/* organizations settings / workspaces — page chrome; grid still uses Tailwind in TSX */
+
+.org-settings-workspaces-page {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-8);
+}
